refactor(processor): extract cache key generation
Move result-cache key construction into a processor helper module so the main processing pipeline no longer owns cache option normalization details.
